chat-assistant/requirements.txt
aiohttp==3.11.9
nats-py==2.9.0
python-dotenv==1.0.0